Windows 7 fallback for libreoffice-help i18n detection
The script now only uses the Windows ≥ 8 specific
cmdlets if the are available. On Windows ≤ 7, it
uses the Get-Culture cmdlet, which is far from
optimal, but currently the workaround I know.
Get-Culture only returns the locale format
settings, not the actual Windows UI language.
For some users, this could pick the wrong
language.