Duplicate some chocolateyInstall.ps1 files for related packages
Previously these were symlinks, but I forgot to integrate them
into the repository. Git for Windows currently does not support
symlinks, so it's necessary to find another solution to avoid
having duplicate scripts. In the meantime this commit add the needed
files for some related packages to not break them.