cyg-get: removed mirror url
It's not required to provide a mirror url, specially if cygwin was already installed (which is a requirement for cyg-get already, anyway).
Alternativelly, we could get a mirror at random (or based on location if possible) from: https://cygwin.com/mirrors.lst  ...but please don't pick the US California mirror like this and make it default.