Update chocolateyInstall.ps1
I noticed a glitch that even power shell didn't catch as being wrong. The  ":\" should only be ":" or the $chromium_string shouldn't start with a '\'
I also added the extra Argument of '--do-not-launch-chrome'. This was removed with all the confusion about the passing of variables to chocolatey.
Tested and works