(core.extension) Get-UninstallRegistryKey: fix PS 2.0 compatibility
Apparently, the trick with accessing a property of all objects in a
collection does not work in PowerShell 2.0, resulting in

WARNING: Found more than 10 bad registry keys. Run command again with
'--verbose --debug' for more info.

The idiomatic, documented and reliable way to do it is via
Select-Object -ExpandProperty, so let's use that.