(vcredist140) improve messages (and fix race condition on older choco)
Seeing the message:

    Installing vcredist140...
    vcredist140 has been installed.

twice might confuse the user. Add architecture suffix to the package name
so that messages are distinct.

Also, this fixes a rare race condition:

Older Chocolatey versions which use fixed local file names
(${pachageName}Install.exe) for downloaded files (and do not attempt to
discover the file name by asking the server), when installing the second
runtime (x64), will attempt to overwrite the vcruntime140Install.exe file.
On Windows 10, this might collide with a system service (DiagTel.exe, if
memory serves) which holds the file open for a brief while after the
installation finishes.
Passing a unique packageName to Install-ChocolateyPackages for each
architecture avoids the problem, because the files will have different
names.

At a high level, this may be interpreted as such that this package
contains two sub-packages.