(GH-621) PowerShell functions - allow ignore and splatting
It is hard to know how future expansion of a PowerShell function is
going to look. When a function is created, it is sometimes set with a
certain number of parameters. When not using `CmdletBinding` other
parameters can be passed and ignored. However, when `CmdletBinding` is
opted into explicitly or implicitly, it doesn't automatically allow for
additional ignored arguments to be passed through. In fact, it out right
errors when you pass more than the arguments that are present. It also
does not allow for splatting with more parameters than are used for a
particular function. This is suboptimal.

Fortunately, for functions with `CmdletBinding`, an additional catch all
parameter can be added to the end to see the other parameters and add
them as part of the function. This allows for future expansion and
allows for older versions of the function to gracefully adapt when
new parameters have been added to the function (they simply ignore
what they don't know about).

Add `IgnoredArguments` to each of the functions in the extensions to
allow for gracefully allowing future expansion and for splatting.

This was seen in Chocolatey's code base in https://github.com/chocolatey/choco/commit/4ee8818d505df99b550041b7dc2c82b3bd909f4f
when it added documentation and needed to enable CmdletBinding to
produce proper docs.