(ruby) Ensure unattended
Ruby now uses RubyInstaller2 as of 2.4, which means a new
way of installing and setting up the DevKit, which is built into
Ruby instead of having to install separately. Unfortunately 
that also means some new changes, as a window pops up 
after the install (thankfully in a non-blocking way) to ask the
user to select how they would like to have the devkit install
proceed. The switch `noridkinstall` prevents that from 
happening. As of any version of Ruby past 2.4.1-1, this will
be the default as well, but it's better to be explicit about 
specifying this behavior.

https://github.com/oneclick/rubyinstaller2/issues/43
https://github.com/oneclick/rubyinstaller2/wiki/FAQ#user-content-silent-install

Fixes #742