Rework updater with two major changes:
1. Use builds from https://github.com/henrypp/chromium which offers
codecs, chrome-sync, and webrtc. With this, chromium is as capable as chrome

2. Specifically follow the "release" versions, mirroring google chrome's
current release version. As chrome is built from chromium, and they use
the same versioning schemes, it follows that the stable version of
chrome is the stable version of chromium. chromium has no official
builds other than a nightly build with reduced feature set.

https://github.com/chocolatey/chocolatey-coreteampackages/issues/736