vcredist140: work around unreliable newer version detection logic
Some versions of vcredist140 do not update the detection values in
the registry correctly (for example, vcredist140 version 14.16.27029
leaves DevDiv\vc\Servicing\14.0\RuntimeMinimum!Version set to 14.16.27024).
This fools the newer version detection logic into thinking the redist
needs to be upgraded. The package proceeds to install the redist, which
fails. The log file states:

[52DC:617C][2019-04-11T10:38:23]i102: Detected related bundle: {f50edb7e-c25e-47b4-bc4f-7ec4a4d256b1}, type: Upgrade, scope: PerMachine, version: 14.16.27029.1, operation: Downgrade
(...)
[52DC:8228][2019-04-11T10:38:23]e000: Error 0x80070666: Cannot install a product when a newer version is installed.

As a workaround, add "a newer version is already installed" to the list
of valid exit codes.