[freecad] manual update
Manual update, automatic update will work again when a version like "0.19.x" will be released.
Until now, the published version is a mix between the major version and build number so to have again a conventionnal versionning with the update script, it needs to manually update this package until a version greater or equal to 0.19 is released by Freecad.