(octave.portable) Update package to launch correctly with current versions
The octave.portable package was not operating correctly.  The desktop icon
and the shims were not launching the package correctly following an
install and there was no icon available for the package.

The issues with the package were caused by updates to the required
parameters and directory structure for the package and an incorrect shim
which did not configure an envrionment appropriately being installed.
To address these issues a number of changes were made:
* Update desktop icon parameters - desktop icon not launching Octave
as `--force-gui` is no longer a supported parameter
* Repoint shim from `octave-cli.exe` to `octave.bat` - launching directly
through `octave-cli.exe` fails to correctly initialize the environment and
leads to octave not functioning correctly
* Update the working directory set for shortcuts to the user profile
rather than the package install directory

In addition to the changes made to ensure the package launched correctly
some functional change were made.  These consisted of:
* Make desktop shortcut installation optional
* Support shortcut installation on the start menu
* Add a shortcut for the Octave CLI
* Add support for shortcut isntallation as a local user
* Document new package parameters