Improve the method of detecting system locale
The current method of detecting system locale is not robust enough. Take "zh-HK" as an example, it should be fallback to "zh_TW", but this package fallback "zh-HK" to "en-US". This commit detects fallback locale by sending `Accept-Language: zh-HK` header to https://www.mozilla.org/ to get Mozilla fallback locale.