(firefox*,thunderbird) Update acquiring locale from from website to use choco helper
This commit updates the helper code for acquiring a supported language code
for languages that have a fallback that would not be considered as we usually
do when checking for two-letter locales.

Due to it being possible to get an invalid culture code even from the website,
there is a absolute fallback to the 'en-US' locale.

fixes #1727