(vcredist140) Fix package update when installer did not change
If the installer ETag did not change, but the package did (due to
versioning decision or a package fix with forced update), the nuspec and
ps1 files still need to be updated with necessary information (resolved
urls, checksums). The Update-OnETagChanged function does not provide
a way to save that information together with the stored ETag, so we need
to obtain it the same way as if the ETag did change.