(tortoisegit) Update package to use latest recommendations
This commit updates the package and package update script
to make use of our latest recommendations by not using globbing
patterns when getting the initial installer.

Additionally, legacy code about the need to select the correct
architecture manually has been removed as this is for much older
versions of Chocolatey CLI than what we support.