(brave) Change to partially embedded package
This commit updates the brave package to be partially embedded.
This means it embeds the 64bit installer, while the 32bit installer is
not included in the package.

This was needed as including both installers has caused the package
to be too large to be uploaded to the Chocolatey Community Repository.

fixes #2158