(umlet) Fix URL to detect install archive
This commit updates the automatic update to fix the URL
used when parsing the URL to use for the embedded archive, as
well as the version.

Additionally, it changes the release notes to be automatically updated
so we won't have to update it multiple places in the future when it is needed.

ref #2187