(maint) Don't include file in email notification
The update_info.xml file can be quite large, and doesn't have directly
usable information, so including it in the email all the time, doesn't
make sense.

In a meeting with Kim, the decision was taken to no longer include this
all the time, and if required, on a per-build basis you can look at the
artifacts creation within the build.