(maint) Update to latest Choco version
This was missed as part of the Chocolatey CLI release process and needs
to be included there going forward.

This commit bumps to the latest version available at the current time.
At the time of this commit, the oldest supported version is still 1.1.0
which is why the choco_version_pr number wasn't changed.