(maint) Update SMTP settings
At some point, not sure when, the sending of emails from the build
stopped working. On investigation, it wasn't clear exactly what the
cause was, but a new app password has been created, encrypted, and
added to the ymo configuration file.  It was also noted that using 465
rather than 587 as the port number was required. so this change was
also made.