(freecad) Adds dependency on Chocolatey >2.2.2 for 7z Package
The package was failing to install. On investigation, there was an issue with the version of 7z being used during unpacking which is solved in later versions bundled with Chocolatey.

This change adds a dependency on Chocolatey to all freecad packages (to simplify the maintenance) and bumps it to 2.2+ in the dev stream.