(cmake) Fix AU version detection
The cmake download page was changed to have both old prerelease versions still available for download
after the stable minor version is released. This causes the AU script to error, as the streams variable already
has the key, so the stable version is not added. This adds a check to fix that.