(apache-httpd) Move package to manual directory
The download of httpd is no longer listed on their download page
which means we are now unable to parse any information about this.

As there is additionally a pending PR to move to a different provider under
a new package name, it is appropriate to move this package to the manual
directory.