(1password7) Split 1password 7.x package to its own
This creates a new package called 1password7 with the code that is
currently used in the 1password package for the 7.x version range.

This also includes updating the package to make use of newer Chocolatey
CLI functionality, and fix an issue with the software start menu
shortcut not showing up.