(1password8) Create package for 1password 8.x
This creates a new package to handle installations of the 1password 8.x
range.

This uses a similar approach as the 1password7 package, but makes us of
the machine wide MSI instead of the User level exe file, as well as
requiring to use the release notes page to get the version as these are
not provided through the normal update service anymore.