(1password) Change to meta package
This updates the 1password package to be a meta package that will have a
dependency on the installation packages of the different 1password major
version packages.

This has been written in a way to attempt to make it easy to add
additional major version packages without having to make any changes to
the meta package manually.