(inkscape) Disable package update
The 32bit installer is missing from the latest version of InkScape, and
the error that is thrown is currently preventing our update gist to be
updated properly.

As such we throw a minimized error message to continue allowing the gist
to be updated, as well as preventing an update until a decision has been
made of whether 32bit support will be dropped or not.