(audacity) Changes Update Logic to use GitHub Releases
The audacity package has failed the last 20+ builds with "Unable to
connect to the remote server". As it seems to work on my local machine,
I suspect the AudacityTeam.org site is blocking our automation.

We were already ending up with the download URLs from the GitHub release
page, so this commit converts the update script to use it for sourcing
the version _and_ the download URLs.