(autohotkey.portable) Simplifies Update Script
The build was seeing "(403) forbidden" when trying to find the latest
version of AutoHotkey. As we were already using GitHub Releases as our
download location, it made sense to convert to using the GH releases API
to source the version information too.

On top of this, this commit simplifies the portable update script by
removing the use of streams.