(poweriso) Update regex for version and urls
The layout of the PowerIso download page has changed, so the previous
regex for finding the information no longer works.  This commit updates
the main regex for finding the latest version number to be
"V (?<version>[\d\.]+)"
and then it extracts the version number.  Following that, rather than
use some partially hard-coded URL's for the downloads, actually do
other regex's to locate the URL's, and then pass them through to AU.