(nexus-repository) Patchfix Install Script with new URL
Sonatype have moved their download URL from fastly to something else.

This breaks existing packages, and as apparently our package cacher did not run against this, all customers and users alike.

As this update is currently in the disabled state (due to previous issues), we cannot simply let it fix itself, and need to manually fix it.

This updates the existing version to a working URL, and uses the package fix notation for anything below the next patch version.