Exit install when pkg_version <= installed_version
The package, as it is in chocolatey, exits without installing whenever the package version is equal to the installed version.

But this means that it's possible for the package to perform a downgrade. For instance, if the user were to run use vscode's own upgrade mechanism then try to upgrade through chocolately.

By having it return when ($version -le (Get-UninstallRegistryKey "$softwareName").DisplayVersion)
This ensures that chocolatey will exit whenever the package version is equal or lower than the installed version. This ensures that it won't perform a downgrade.